Disable support for Node versions < 4.0 without syntax errors.
Create a layer of abstraction to not expose newer node.js features that'd
cause syntax errors when require'd in older versions of node.js.

Also exported a ModuleSupported field so that the app does not need to be
aware of what platforms and Node versions are supported.
